requirements.txt
# --- Core FastAPI & Web Server ---
fastapi==0.128.0
uvicorn[standard]==0.40.0
starlette==0.50.0
python-multipart==0.0.21

# --- Pydantic ---
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydantic-extra-types==2.11.0

# --- Database ---
PyMySQL==1.1.1
SQLAlchemy==2.0.45

# --- AI & Machine Learning ---
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.81
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
sentence-transformers==3.0.1
huggingface-hub==0.36.0
transformers==4.45.2
torch==2.9.1
tokenizers==0.20.3
safetensors==0.7.0
tqdm==4.66.5

# --- Vector Database ---
pinecone==4.1.0
pinecone-client==4.1.0

# --- Environment & Utilities ---
python-dotenv==1.0.1
requests==2.32.3
langdetect==1.0.9
aiohttp==3.13.3
httpx==0.28.1
anyio==4.12.0

numpy==2.1.0